Déplace la route "/visiteGuidee/:idEtape" vers le routeur dédié
