+++ src/mss.js
@@ -1,7 +1,6 @@
 const cookieSession = require('cookie-session');
 const cookieParser = require('cookie-parser');
 const express = require('express');
-const { decode } = require('html-entities');
 const {
   CACHE_CONTROL_FICHIERS_STATIQUES,
   DUREE_SESSION,
@@ -14,9 +13,6 @@ const {
   routesNonConnecteApiBibliotheques,
 } = require('./routes/nonConnecte/routesNonConnecteApiBibliotheques');
 const routesNonConnecteApiStyles = require('./routes/nonConnecte/routesNonConnecteApiStyles');
-const InformationsHomologation = require('./modeles/informationsHomologation');
-const Service = require('./modeles/service');
-const Utilisateur = require('./modeles/utilisateur');
 const routesNonConnectePage = require('./routes/nonConnecte/routesNonConnectePage');
 const routesConnectePage = require('./routes/connecte/routesConnectePage');
 
@@ -76,7 +72,15 @@ const creeServeur = (
   app.set('views', './src/vues');
 
   app.use('', routesNonConnectePage({ depotDonnees, middleware, referentiel }));
-  app.use('', routesConnectePage({ depotDonnees, middleware, referentiel }));
+  app.use(
+    '',
+    routesConnectePage({
+      depotDonnees,
+      middleware,
+      moteurRegles,
+      referentiel,
+    })
+  );
 
   app.use(
     '/api',
@@ -122,69 +126,6 @@ const creeServeur = (
       adaptateurGestionErreur,
       adaptateurHorloge,
     })
-  );
-
-  app.get(
-    '/visiteGuidee/:idEtape',
-    middleware.verificationJWT,
-    middleware.chargePreferencesUtilisateur,
-    middleware.chargeEtatVisiteGuidee,
-    (requete, reponse) => {
-      const utilisateurVisiteGuidee = new Utilisateur({
-        email: 'visite-guidee@cyber.gouv.fr',
-      });
-      const service = Service.creePourUnUtilisateur(utilisateurVisiteGuidee);
-      service.id = 'ID-SERVICE-VISITE-GUIDEE';
-
-      const { idEtape } = requete.params;
-      reponse.locals.etatVisiteGuidee = {
-        ...reponse.locals.etatVisiteGuidee,
-        etapeCourante: idEtape.toUpperCase(),
-      };
-      reponse.locals.autorisationsService = {
-        DECRIRE: { estMasque: false },
-        SECURISER: { estMasque: false, estLectureSeule: false },
-        HOMOLOGUER: { estMasque: false },
-        RISQUES: { estMasque: false },
-        CONTACTS: { estMasque: false },
-        peutHomologuer: false,
-      };
-
-      if (idEtape === 'decrire') {
-        reponse.render('service/creation', {
-          InformationsHomologation,
-          referentiel,
-          service,
-          etapeActive: 'descriptionService',
-          departements: referentiel.departements(),
-        });
-      } else if (idEtape === 'securiser') {
-        const mesures = moteurRegles.mesures(service.descriptionService);
-        const pourcentageProgression = 80;
-
-        service.indiceCyber = () => ({ total: 4.3 });
-        reponse.render('service/mesures', {
-          InformationsHomologation,
-          referentiel,
-          service,
-          etapeActive: 'mesures',
-          pourcentageProgression,
-          mesures,
-        });
-      } else if (idEtape === 'homologuer') {
-        reponse.render('service/dossiers', {
-          InformationsHomologation,
-          decode,
-          service,
-          etapeActive: 'dossiers',
-          premiereEtapeParcours: referentiel.premiereEtapeParcours(),
-          peutVoirTamponHomologation: true,
-          referentiel,
-        });
-      } else if (idEtape === 'piloter') {
-        reponse.render('tableauDeBord');
-      }
-    }
   );
 
   app.use(
